requirements.txt
mysql-connector-python==9.0.0
PySide6==6.8.1.1